Diffusion Approximation

Infinite

LightPropagation.fluence_DA_inf_CW — Function

fluence_DA_inf_CW(ρ, μa, μsp)

Compute the steady-state fluence in an infinite medium.

Arguments

ρ: source-detector separation (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

Examples

julia> fluence_DA_inf_CW(1.0, 0.1, 10.0)

LightPropagation.fluence_DA_inf_TD — Function

fluence_DA_inf_TD(t, ρ, μa, μsp; nmed = 1.0)

Compute the time-domain fluence in an infinite medium with Eqn. 3 of Patterson. et al. 1989.

Arguments

t: the time vector (ns).

ρ: source-detector separation (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

n_med::Float64: medium's index of refraction

Examples

julia> fluence_DA_inf_TD(0.1:0.5:5.0, 1.0, 0.1, 10.0, n_med = 1.4)

Semi-infinite

LightPropagation.fluence_DA_semiinf_CW — Function

fluence_DA_semiinf_CW(ρ, μa, μsp; n_ext = 1.0, n_med = 1.0, z = 0.0)

Compute the steady-state fluence in a semi-infinite geometry according to Eqn. 3 of Kienle 1997.

Arguments

ρ: the source detector separation (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

z: the z-depth orthogonal from the boundary (cm)

Examples

julia> fluence_DA_semiinf_CW(1.0, 0.1, 10.0, n_ext = 1.0, n_med = 1.0, z = 0.0)

LightPropagation.fluence_DA_semiinf_TD — Function

fluence_DA_semiinf_TD(t, ρ, μa, μsp; n_ext = 1.0, n_med = 1.0, z = 0.0)

Compute the time-domain fluence in a semi-infinite medium (Eqn. 33 Contini).

Arguments

t: the time vector (ns).

ρ: the source detector separation (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

z: the z-depth in medium

Examples

julia> fluence_DA_semiinf_TD(0.1:0.1:1.0, 1.0, 0.1, 10.0, n_ext = 1.0, n_med = 1.0, z = 0.0)

Slab

LightPropagation.fluence_DA_slab_CW — Function

fluence_DA_slab_CW(ρ, μa, μsp; n_ext = 1.0, n_med = 1.0, s = 2.0, z = 0.0, xs = 10)

Compute the steady-state fluence from a slab geometry (x, y -> inf, z -> finite).

Arguments

ρ: the source detector separation (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

s: the thickness (z-depth) of the slab (cm)

z: the z-depth within slab (cm)

xs: the number of sources to compute in the series

Examples

julia> fluence_DA_slab_CW(1.0, 0.1, 10.0, n_ext = 1.0, n_med = 1.0, s = 1.0, z = 0.0)

LightPropagation.fluence_DA_slab_TD — Function

fluence_DA_slab_TD(t, ρ, μa, μsp; n_ext = 1.0, n_med = 1.0, s = 1.0, z = 0.0, xs = 10)

Compute the time-domain fluence from a slab geometry (x, y -> inf, z -> finite).

Arguments

t: the time vector (ns).

ρ: the source detector separation (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

s: the thickness (z-depth) of the slab (cm)

z: the z-depth coordinate (cm)

xs: the number of sources to compute in the series

Examples

julia> fluence_DA_slab_TD(0.1, 1.0, 0.1, 10.0, s = 40.0)

Parallelepiped

LightPropagation.fluence_DA_paralpip_CW — Function

fluence_DA_paralpip_CW(μa, μsp; n_ext = 1.0, n_med = 1.0, rd = [4.0, 5.0, 0.0], rs = [5.0, 5.0], L = [10.0, 10.0, 10.0], xs = 10)

Compute the steady-state fluence in a parallelepiped [lx, ly, lz].

Arguments

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

rd: target location within medium [x,y,z] with origin x,y = 0 at corner of parallelepiped; z assumed to be 0

rs: the location of the source [x,y] with origin x,y = 0 at corner of parallelepiped; z assumed to be 0

L: the dimensions [lx, ly, lz] of the parallelepiped

xs: the number of sources to compute in the series

Examples

julia> fluence_DA_paralpip_CW(0.1, 10.0, n_ext = 1.0, n_med = 1.0, rd = [24.0, 25.0, 0.0], rs = [25.0,25.0], L = [50.0,50.0,50.0], xs = 20)

LightPropagation.fluence_DA_paralpip_TD — Function

fluence_DA_paralpip_TD(t, μa, μsp; n_ext = 1.0, n_med = 1.0, rd = [4.0, 5.0, 0.0], rs = [5.0, 5.0], L = [10.0, 10.0, 10.0], xs = 10)

Compute the time-domain fluence in a parallelepiped [lx, ly, lz].

Arguments

t: the time vector (ns).

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

rd: target location within medium [x,y,z] with origin x,y = 0 at corner of parallelepiped; z assumed to be 0

rs: the location of the source [x,y] with origin x,y = 0 at corner of parallelepiped; z assumed to be 0

L: the dimensions [lx, ly, lz] of the parallelepiped

xs: the number of sources to compute in the series

Examples

julia> fluence_DA_paralpip_TD(0.5, 0.1, 10.0, n_ext = 1.0, n_med = 1.0, rd = [24.0, 25.0, 0.0], rs = [25.0,25.0], L = [50.0,50.0,50.0], xs = 20)