Diffusion Approximation

Infinite

LightPropagation.fluence_DA_inf_CW — Function

fluence_DA_inf_CW(ρ, μa, μsp)

Compute the steady-state fluence in an infinite medium. This is an unsafe implementation and will not check parameters. ρ and μsp should be > 0.0 and μa >= 0.0.

Arguments

ρ: source-detector separation (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

Examples

julia> fluence_DA_inf_CW(1.0, 0.1, 10.0)

fluence_DA_inf_CW(data::DiffusionParameters)

Wrapper to fluence_DA_inf_CW(ρ, μa, μsp)

Examples

julia> data = DAinf_params(ρ = 1.0) # use structure to generate inputs
julia> fluence_DA_inf_CW(data) # then call the function

LightPropagation.fluence_DA_inf_TD — Function

fluence_DA_inf_TD(t, ρ, μa, μsp; n_med = 1.0)

Compute the time-domain fluence in an infinite medium with Eqn. 3 of Patterson. et al. 1989. This is an unsafe implementation and will not check parameters. t, ρ and μsp should be > 0.0 and μa >= 0.0.

Arguments

t: the time vector (ns).

ρ: source-detector separation (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

Keyword Arguments

n_med= 1.0: medium's index of refraction

Examples

julia> fluence_DA_inf_TD(0.1:0.5:5.0, 1.0, 0.1, 10.0, n_med = 1.4)

fluence_DA_inf_TD(data::DiffusionParameters)

Wrapper to fluence_DA_inf_TD(t, ρ, μa, μsp; n_med = 1.0).

Examples

julia> data = DAinf_params(t = 0.1:0.05:2.0) # use structure to generate inputs
julia> fluence_DA_inf_TD(data) # then call the function

Semi-infinite

LightPropagation.fluence_DA_semiinf_CW — Function

fluence_DA_semiinf_CW(ρ, μa, μsp; n_ext = 1.0, n_med = 1.0, z = 0.0)

Compute the steady-state fluence in a semi-infinite geometry.

Arguments

ρ: the source detector separation (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

Keyword arguments

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

z: the z-depth orthogonal from the boundary (cm)

Examples

julia> fluence_DA_semiinf_CW(1.0, 0.1, 10.0)
julia> fluence_DA_semiinf_CW(1.0, 0.1, 10.0, n_ext = 1.0, n_med = 1.0, z = 0.0)

fluence_DA_semiinf_CW(data::DiffusionParameters)

Wrapper to fluence_DA_semiinf_CW(ρ, μa, μsp; n_ext = 1.0, n_med = 1.0, z = 0.0...)

Examples

julia> data = DAsemiinf_params(ρ = 1.0) # use structure to generate inputs
julia> fluence_DA_semiinf_CW(data) # then call the function

LightPropagation.fluence_DA_semiinf_TD — Function

fluence_DA_semiinf_TD(t, ρ, μa, μsp; n_ext = 1.0, n_med = 1.0, z = 0.0)

Compute the time-domain fluence in a semi-infinite medium.

Arguments

t: the time vector (ns).

ρ: the source detector separation (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

Optional arguments

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

z: the z-depth orthogonal from the boundary (cm)

Examples

julia> fluence_DA_semiinf_TD(0.1:0.1:2.0, 1.0, 0.1, 10.0)
julia> fluence_DA_semiinf_TD(0.2:0.2:1.0, 1.0, 0.1, 10.0, n_ext = 1.0, n_med = 1.0, z = 0.0)

fluence_DA_semiinf_TD(t, data::DiffusionParameters)

Wrapper to fluence_DA_semiinf_TD(t, ρ, μa, μsp; n_ext = 1.0, n_med = 1.0, z = 0.0)

Examples

julia> data = DAsemiinf_params() # use structure to generate inputs
julia> fluence_DA_semiinf_TD(data) # then call the function

LightPropagation.flux_DA_semiinf_CW — Function

flux_DA_semiinf_CW(ρ, μa, μsp; n_ext = 1.0, n_med = 1.0)

Compute the steady-state flux (D*∂ϕ(ρ)/∂z @ z = 0) from a semi-infinite medium.

Arguments

ρ: the source detector separation (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

ndet: the boundary's index of refraction (air or detector)

nmed: the sample medium's index of refraction

Examples

julia> flux_DA_semiinf_CW(1.0, 0.1, 10.0, n_ext = 1.0, n_med = 1.0)

flux_DA_semiinf_CW(data::DiffusionParameters)

Wrapper to flux_DA_semiinf_CW(ρ, μa, μsp; n_ext = 1.0, n_med = 1.0)

Examples

julia> data = DAsemiinf_params(ρ = 1.0) # use structure to generate inputs
julia> flux_DA_semiinf_CW(data) # then call the function

LightPropagation.flux_DA_semiinf_TD — Function

flux_DA_semiinf_TD(t, ρ, μa, μsp; n_ext = 1.0, n_med = 1.0)

Compute the time-domain flux (D*∂ϕ(t)/∂z @ z = 0) from a semi-infinite medium from Eqn. 36 Contini 97 (m = 0).

Arguments

t: the time vector (ns).

ρ: the source detector separation (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

Keyword arguments

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

Examples

julia> flux_DA_semiinf_TD(0.1:0.1:2.0, 1.0, 0.1, 10.0, n_ext = 1.0, n_med = 1.0)

flux_DA_semiinf_TD(t, data::DiffusionParameters)

Wrapper to flux_DA_semiinf_TD(t, ρ, μa, μsp; n_ext = 1.0, n_med = 1.0, z = 0.0)

Examples

julia> data = DAsemiinf_params() # use structure to generate inputs
julia> flux_DA_semiinf_TD(data) # then call the function

Slab

LightPropagation.fluence_DA_slab_CW — Function

fluence_DA_slab_CW(ρ, μa, μsp; n_ext = 1.0, n_med = 1.0, s = 2.0, z = 0.0, xs = 10)

Compute the steady-state fluence from a slab geometry (x, y -> inf, z -> finite). The fluence is calculated from an infinite summation that converges rather rapidly. Generally only 5 terms are needed (xs = 5), but more terms will be required if ρ >> s.

Arguments

ρ: the source detector separation (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

Optional arguments

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

s: the thickness (z-depth) of the slab (cm)

z: the z-depth within slab (cm)

xs: the number of sources to compute in the series

Examples

julia> fluence_DA_slab_CW(1.0, 0.1, 10.0, n_ext = 1.0, n_med = 1.0, s = 1.0, z = 0.0)

fluence_DA_slab_CW(data::DiffusionParameters)

Wrapper to fluence_DA_slab_CW(ρ, μa, μsp; n_ext = 1.0, n_med = 1.0, z = 0.0, s = 1.0, xs = 10)

Examples

julia> data = DAslab(ρ = 1.0, s = 3.0) # use structure to generate inputs
julia> fluence_DA_slab_CW(data) # then call the function

LightPropagation.fluence_DA_slab_TD — Function

fluence_DA_slab_TD(t, ρ, μa, μsp; n_ext = 1.0, n_med = 1.0, s = 1.0, z = 0.0, xs = 10)

Compute the time-domain fluence from a slab geometry (x, y -> inf, z -> finite). The fluence is calculated from an infinite summation that converges rather rapidly. Generally only 5 terms are needed (xs = 5), but more terms will be required if ρ >> s. At long time values (low fluence values), the solution struggles with limited precision. Using double floats or BigFloats is advised at these times.

Arguments

t: the time vector (ns).

ρ: the source detector separation (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

Optional arguments

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

s: the thickness (z-depth) of the slab (cm)

z: the z-depth coordinate (cm)

xs: the number of sources to compute in the series

Examples

julia> fluence_DA_slab_TD(0.1, 1.0, 0.1, 10.0, s = 40.0)

fluence_DA_slab_TD(t, data::DiffusionParameters)

Wrapper to fluence_DA_slab_TD(t, ρ, μa, μsp; n_ext = 1.0, n_med = 1.0, s = 1.0, z = 0.0, xs = 10)

Examples

julia> data = DAslab() # use structure to generate inputs
julia> fluence_DA_slab_TD(data) # then call the function

LightPropagation.flux_DA_slab_CW — Function

fluence_DA_slab_CW(ρ, μa, μsp; n_ext = 1.0, n_med = 1.0, s = 2.0, z = 0.0, xs = 10)

Compute the steady-state flux, D∂ϕ(ρ)/∂z for z = 0 and -D∂ϕ(ρ)/∂z for z = s from a slab geometry (x, y -> inf, z -> finite). If z != 0 or s will default to compute D*∂ϕ(ρ)/∂z for the z given.

Arguments

ρ: the source detector separation (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

Optional arguments

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

s: the thickness (z-depth) of the slab (cm)

z: the z-depth within slab (cm)

xs: the number of sources to compute in the series

Examples

julia> flux_DA_slab_CW(1.0, 0.1, 10.0, n_ext = 1.0, n_med = 1.0, s = 2.0, z = 0.0)

flux_DA_slab_CW(data::DiffusionParameters)

Wrapper to flux_DA_slab_CW(ρ, μa, μsp; n_ext = 1.0, n_med = 1.0)

Examples

julia> data = DAslab(ρ = 1.0) # use structure to generate inputs
julia> flux_DA_slab_CW(data) # then call the function

LightPropagation.flux_DA_slab_TD — Function

flux_DA_slab_TD(t, ρ, μa, μsp; n_ext = 1.0, n_med = 1.0, s = 1.0, z = 0.0, xs = 15)

Compute the time-domain flux, D*∂ϕ(t)/∂z for z = 0 and -D*∂ϕ(t)/∂z for z = s from a slab geometry (x, y -> inf, z -> finite).
If z != 0 or s will default to compute D*∂ϕ(t)/∂z for the z given.

Arguments

t: the time vector (ns).

ρ: the source detector separation (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

Keyword arguments

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

s: the thickness (z-depth) of the slab (cm)

z: the z-depth coordinate (cm) must be equal

xs: the number of sources to compute in the series

Examples

julia> flux_DA_slab_TD(1.0, 1.0, 0.1, 10.0, n_ext = 1.0, n_med = 1.0, z = 0.0, s = 1.0)

flux_DA_slab_TD(t, data::DiffusionParameters)

Wrapper to flux_DA_slab_TD(t, ρ, μa, μsp; n_ext = 1.0, n_med = 1.0, z = 0.0, s = 1.0, xs = 15)

Examples

julia> data = DAslab(ρ = 1.0) # use structure to generate inputs
julia> flux_DA_slab_TD(data) # then call the function

Parallelepiped

LightPropagation.fluence_DA_paralpip_CW — Function

fluence_DA_paralpip_CW(μa, μsp; n_ext = 1.0, n_med = 1.0, rd = [4.0, 5.0, 0.0], rs = [5.0, 5.0], L = [10.0, 10.0, 10.0], xs = 10)

Compute the steady-state fluence in a parallelepiped [lx, ly, lz].

Arguments

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

Optional arguments

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

rd: target location within medium [x,y,z] with origin x,y = 0 at corner of parallelepiped; z assumed to be 0

rs: the location of the source [x,y] with origin x,y = 0 at corner of parallelepiped; z assumed to be 0

L: the dimensions [lx, ly, lz] of the parallelepiped

xs: the number of sources to compute in the series

Examples

julia> fluence_DA_paralpip_CW(0.1, 10.0, n_ext = 1.0, n_med = 1.0, rd = [24.0, 25.0, 0.0], rs = [25.0,25.0], L = [50.0,50.0,50.0], xs = 20)

LightPropagation.fluence_DA_paralpip_TD — Function

fluence_DA_paralpip_TD(t, μa, μsp; n_ext = 1.0, n_med = 1.0, rd = [4.0, 5.0, 0.0], rs = [5.0, 5.0], L = [10.0, 10.0, 10.0], xs = 10)

Compute the time-domain fluence in a parallelepiped [lx, ly, lz].

Arguments

t: the time vector (ns).

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

Optional arguments

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

rd: target location within medium [x,y,z] with origin x,y = 0 at corner of parallelepiped; z assumed to be 0

rs: the location of the source [x,y] with origin x,y = 0 at corner of parallelepiped; z assumed to be 0

L: the dimensions [lx, ly, lz] of the parallelepiped

xs: the number of sources to compute in the series

Examples

julia> fluence_DA_paralpip_TD(0.5, 0.1, 10.0, n_ext = 1.0, n_med = 1.0, rd = [24.0, 25.0, 0.0], rs = [25.0,25.0], L = [50.0,50.0,50.0], xs = 20)

Layered Cylinder

LightPropagation.fluence_DA_Nlay_cylinder_CW — Function

fluence_DA_Nlay_cylinder_CW(ρ, μa, μsp, n_ext, n_med, l, a, z, besselroots)

Compute the steady-state fluence in an N-layered cylinder. Source is assumed to be located on the top middle of the cylinder. No checks on arguments are made except enforcing that the source is located in the top layer. ρ, μa, and z should be >= 0.0. μsp, next, nmed, a should be > 0.0. Each layer in l should be > 0.0. In general, ρ should always be < a. All arguments should have the same types and the length of μsp, μa, nmed, l should be equal. It is advised to use the Nlayercylinder() structure format for inputs.

Arguments

ρ: source-detector separation in cylindrical coordinates (distance from middle z-axis of cylinder) (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

l: layer thicknesses (cm)

a: cylinder radius (cm)

z: source depth within cylinder

besselroots: roots of bessel function of first kind zero order J0(x) = 0

Examples

julia> fluence_DA_Nlay_cylinder_CW(1.0, [0.1, 0.1], [10.0, 10.0], 1.0, [1.0, 1.0], [4.5, 4.5], 10.0, 0.0, besselroots)

fluence_DA_Nlay_cylinder_CW(data)

Wrapper to fluenceDANlaycylinderCW(ρ, μa, μsp, next, nmed, l, a, z, besselroots) with inputs given as a structure (data).

Examples

julia> data = Nlayercylinder(a = 10.0, l = [1.0, 1.0, 1.0, 2.0], z = 5.0) julia> `fluenceDANlaycylinder_CW(data)`

LightPropagation.fluence_DA_Nlay_cylinder_TD — Function

fluence_DA_Nlay_cylinder_TD(t, ρ, μa, μsp, n_ext, n_med, l, a, z, besselroots; N = 24, ILT = hyper_fixed)

Compute the time-domain fluence in an N-layered cylinder. Source is assumed to be located on the top middle of the cylinder. See notes on fluenceDANlaycylinderCW for input checks with the initial constraint that all values of t should be > 0.0 and ordered least to greatest. Do not have any value of t be zero or negative.

It is best to use hyper_fixed as the inverse laplace transform if t consists of many time points. Utilize hyperbola for a single time point. The value of N should be proportional to the dynamic range of the time-domain signal needed. For later times you will need a larger N. The lowest fluence value you can compute will be no less than the machine precision you utilize. Float64 values are limited to fluence > ~2.2-16.

Arguments

t: time point or vector (ns)

ρ: source-detector separation in cylindrical coordinates (distance from middle z-axis of cylinder) (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

l: layer thicknesses (cm)

a: cylinder radius (cm)

z: source depth within cylinder

besselroots: roots of bessel function of first kind zero order J0(x) = 0

N: number of Hankel-Laplace calculations

ILT: inverse laplace transform function

Examples

julia> fluence_DA_Nlay_cylinder_TD(0.1:0.1:2.0, [0.1, 0.1], [10.0, 10.0], 1.0, [1.0, 1.0], [4.5, 4.5], 10.0, 0.0, besselroots)

fluence_DA_Nlay_cylinder_TD(t, data; N = 24)

Wrapper to fluenceDANlaycylinderTD(t, ρ, μa, μsp, next, nmed, l, a, z, besselroots; N = 24, ILT = hyper_fixed) with inputs given as a structure (data).

Examples

julia> data = Nlayercylinder(a = 10.0, l = [1.0, 1.0, 1.0, 2.0], z = 5.0) julia> `fluenceDANlaycylinder_TD(0.5:1.0:2.5, data)`

LightPropagation.flux_DA_Nlay_cylinder_CW — Function

flux_DA_Nlay_cylinder_CW(ρ, μa, μsp, n_ext, n_med, l, a, z, besselroots)

Compute the steady-state flux using Fick's law D[1]∂ϕ(ρ)/∂z for z = 0 (reflectance) and -D[end]∂ϕ(ρ)/∂z for z = sum(l) (transmittance) in an N-layered cylinder. Source is assumed to be located on the top middle of the cylinder. z must be equal to 0 or the total length sum(l) of cylinder.

Arguments

ρ: source-detector separation in cylindrical coordinates (distance from middle z-axis of cylinder) (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

l: layer thicknesses (cm)

a: cylinder radius (cm)

z: source depth within cylinder: must be equal to 0 or sum(l)

besselroots: roots of bessel function of first kind zero order J0(x) = 0

Examples

julia> flux_DA_Nlay_cylinder_CW(1.0, [0.1, 0.1], [10.0, 10.0], 1.0, [1.0, 1.0], [4.5, 4.5], 10.0, 0.0, besselroots)

flux_DA_Nlay_cylinder_CW(data, besselroots)

Wrapper to fluxDANlaycylinderCW(ρ, μa, μsp, next, nmed, l, a, z, besselroots) with inputs given as a structure (data).

Examples

julia> data = Nlayercylinder(a = 10.0, l = [1.0, 1.0, 1.0, 2.0], z = 5.0) julia> `fluxDANlaycylinder_CW(data, besselroots)`

LightPropagation.flux_DA_Nlay_cylinder_TD — Function

flux_DA_Nlay_cylinder_TD(t, ρ, μa, μsp, n_ext, n_med, l, a, z, besselroots; N = 24)

Compute the time-domain flux using Fick's law D[1]∂ϕ(ρ, t)/∂z for z = 0 (reflectance) and -D[end]∂ϕ(ρ, t)/∂z for z = sum(l) (transmittance) in an N-layered cylinder. Source is assumed to be located on the top middle of the cylinder. z must be equal to 0 or the total length sum(l) of cylinder. Uses the hyperbola contour if t is an AbstractFloat and the fixed hyperbola contour if t is an AbstractVector. ∂ϕ(ρ, t)/∂z is calculated using forward mode auto-differentiation with ForwardDiff.jl

Arguments

t: time point or vector (ns)

ρ: source-detector separation in cylindrical coordinates (distance from middle z-axis of cylinder) (cm⁻¹)

μa: absorption coefficient (cm⁻¹)

μsp: reduced scattering coefficient (cm⁻¹)

n_ext: the boundary's index of refraction (air or detector)

n_med: the sample medium's index of refraction

l: layer thicknesses (cm)

a: cylinder radius (cm)

z: source depth within cylinder

besselroots: roots of bessel function of first kind zero order J0(x) = 0

N: number of Hankel-Laplace calculations

Examples

julia> fluence_DA_Nlay_cylinder_TD(0.1:0.1:2.0, [0.1, 0.1], [10.0, 10.0], 1.0, [1.0, 1.0], [4.5, 4.5], 10.0, 0.0, besselroots)

flux_DA_Nlay_cylinder_TD(t, data; N = 24)

Wrapper to fluxDANlaycylinderTD(t, ρ, μa, μsp, next, nmed, l, a, z, besselroots; N = 24) with inputs given as a structure (data).

Examples

julia> data = Nlayercylinder(a = 10.0, l = [1.0, 1.0, 1.0, 2.0], z = 5.0) julia> `fluxDANlaycylinder_TD(0.5:1.0:2.5, data)`

DCS

Missing docstring.

Missing docstring for DAsemiinf_DCS. Check Documenter's build log for details.

Missing docstring.

Missing docstring for g2_DA_semiinf_CW. Check Documenter's build log for details.

LightPropagation.Nlayer_cylinder_DCS — Type

NlayercylinderDCS

Provides default parameters to use to compute the autocorrelation function g2 in a N-layered cylinder. The arguments defined with a Vector should be the same length and same type.

Keyword Arguments

ρ: source-detector separation (cm⁻¹)

μa::Vector{T}: absorption coefficient (cm⁻¹)

μsp::Vector{T}: reduced scattering coefficient (cm⁻¹)

n_med::Vector{T}: medium's index of refraction

n_ext: external medium's index of refraction (air or detector)

l: the thicknesses of each layer (cm)

z: the z-depth orthogonal from the boundary (cm)

a: the radius of the cylinder (cm)

β: constant in Siegert relation dependent on collection optics

BFi::Vector{T}: Blood flow index ~αDb (cm²/s)

λ: wavelength (nm)

bessels::AbstractVector{T}: roots of J0

Examples

julia> data = Nlayer_cylinder_DCS() # return default parameters
julia> data = Nlayer_cylinder_DCS(ρ = 1.5) # return ρ = 1.5 with the rest of the parameters given by defaults

# we can now define our correlation times τ
julia> τ = 10 .^(range(-10,stop=0,length=250))
julia> g2_DA_Nlay_cylinder_CW(τ, data)

Missing docstring.

Missing docstring for g2_DA_Nlay_cylinder_CW. Check Documenter's build log for details.